% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utility-data.R
\docType{data}
\name{curve.palettes}
\alias{curve.palettes}
\title{Curve Palettes}
\format{
An object of class \code{character} of length 10.
}
\usage{
data(curve.palettes)
}
\description{
A vector of colors for performance curve
}
\examples{
data(curve.palettes)
\dontrun{
curve.palettes
}
}
\keyword{datasets}
