% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.plot.R
\name{dist.plot}
\alias{dist.plot}
\title{Distribution Plots of QDRS}
\usage{
dist.plot(
  disease,
  output.date = NULL,
  score.mat,
  score.name,
  group,
  cutoff = c(seq(0.05, 0.45, by = 0.05), seq(0.5, 1, by = 0.01)),
  unknown.show = FALSE
)
}
\arguments{
\item{disease}{A string of disease name.}

\item{output.date}{An output date. The default is system date (today).}

\item{score.mat}{A resulting score matrix with multiple types of scores.}

\item{score.name}{The score of interest.}

\item{group}{A grouping vector.}

\item{cutoff}{A vector of cutoff points to set the bins of prevalence plot. The default vector has 60 bins.}

\item{unknown.show}{A logical value indicates whethe unknowns should be used for plotting.}
}
\value{
It produces a pdf file of plots and returns the list of three distribution plots.
}
\description{
Generate a pdf file with a density plot, a boxplot and a prevalence plot for the resulting scores. These three plots show the distribution of scores by disease status for visualization of risk stratification.
}
\examples{
\dontrun{
group1 = example.scores$group
group1[group1 != "Control"] = "Case"
set.seed(830)
na.ind = sample(1:length(group1), size = 1000)
group1[na.ind] = NA
res = dist.plot(disease = "Disease",
  score.mat = example.scores$score.mat,
  score.name = "LPC",
  group = group1,
  unknown.show = T)
}
}
