% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHR-data.R
\docType{data}
\name{EHR}
\alias{EHR}
\title{An example data set}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(EHR)
}
\description{
It contains \code{sample.set} a set of 95 features for 5,000 subjects, \code{sample.group} a vector of group with two levels "Case" and "Control", and \code{training} a logical vector that indicates the usage for training.
}
\examples{
data(EHR)
\dontrun{
# check the number of cases and controls
table(EHR$sample.group)
# check the prevalence of features
colMeans(EHR$sample.set)
}
}
\keyword{datasets}
