% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{eigen.score}
\alias{eigen.score}
\title{Eigen Score}
\usage{
eigen.score(X, training)
}
\arguments{
\item{X}{The original data set that include training and test sets. It should be a matrix of numbers.}

\item{training}{A logical or index vector to indicate whether the subject belongs to the training set.}
}
\value{
It returns a list of following components:
\item{weights}{The Eigen weights for input features.}
\item{scores}{The resulting Eigen scores for the whole set.}
}
\description{
Compute Eigen weights and scores. This method is unsupervised, and assigns weights that are proportional to the balanced accuracies (the average between the sensitivity and the specificity) of the input feature.
}
\examples{
\dontrun{
data(EHR)
res1 <- eigen.score(EHR$sample.set, EHR$training)
}
}
\references{
Iuliana Ionita-Laza, Kenneth McCallum, Bin Xu, and Joseph D Buxbaum. A spectral approach integrating functional genomic annotations for coding and noncoding variants. Nature genetics, 48(2):214, 2016.
}
