% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC.R
\name{LPC}
\alias{LPC}
\title{Linear Combination of Principal Components}
\usage{
LPC(X, group, training, scale = TRUE)
}
\arguments{
\item{X}{The original data set that include training and test sets. It should be a matrix of numbers.}

\item{group}{A vector that indicate cases ("Case") and controls ("Control"). \code{NA} is allowed, and means that the observation is not used in individual PC sign determination.}

\item{training}{A logical or index vector to indicate whether the subject belongs to the training set.}

\item{scale}{A logical value to indicate whether the input features need to be scaled.}
}
\value{
It returns a list of following components:
\item{lpc.n}{The number of significant PCs (eigenvalues) suggested by Tracy-Widom test, a vector of the sign of individual PC.}
\item{pc.sign}{A vector of the sign of individual PC.}
\item{weights}{The LPC weights for input features.}
\item{scores}{The resulting LPC scores for the whole set.}
}
\description{
Computes LPC weights and scores based on an almost unsupervised method that combines multiple PCs and only requires weak labels to help select the signs of individual PCs.
}
\examples{
\dontrun{
data(EHR)
res1 <- LPC(X = EHR$sample.set, group = EHR$sample.group, training = EHR$training)
}
}
