% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVS.R
\name{LVS.fit}
\alias{LVS.fit}
\title{Fit a Latent Variable Model.}
\usage{
LVS.fit(X, starting.choice = "random", p.seed = 3080)
}
\arguments{
\item{X}{The training data set. It should be a matrix of binary numbers.}

\item{starting.choice}{Starting values can be generated by fitting model without latent variables, and applying factorial analysis to residuals to get starting values for latent variables and their coefficients (starting.choice = "res"). Another options are to use zeros as a starting values (starting.choice = "zero") or initialize starting values for latent variables with (n x 1) matrix. Defaults to "random".}

\item{p.seed}{A random seed for model fitting.}
}
\value{
An object of class "gllvm" includes the following components.
}
\description{
\code{LVS.fit} computes a generalized lineatr latent variable model for multivariate bernoulli data, which assumes one latent variable.
}
