% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVS.R
\name{LVS.score}
\alias{LVS.score}
\title{Latent Variable Score}
\usage{
LVS.score(
  X,
  Y = NULL,
  family = "binomial",
  starting.choice = "random",
  p.seed = 3080
)
}
\arguments{
\item{X}{A training data set. It should be a matrix of numbers.}

\item{Y}{A new data set. It should be a matrix of numbers. Default is NULL, then X will be used in prediction.}

\item{family}{distribution function for responses. Options are poisson(link = "log"), "negative.binomial" (with log link), binomial(link = "probit") , gaussian(link = "identity"), "gamma" (with log link), and "ordinal".}

\item{starting.choice}{Starting values can be generated by fitting model without latent variables, and applying factorial analysis to residuals to get starting values for latent variables and their coefficients (starting.choice = "res"). Another options are to use zeros as a starting values (starting.choice = "zero") or initialize starting values for latent variables with (n x 1) matrix. Defaults to "random".}

\item{p.seed}{A random seed for model fitting.}
}
\value{
lvs The resulting LVSs for the new set Y.
}
\description{
Computes Latent Variable Scores based on an unsupervised multivariate mixed model framework for multivariate bernoulli data that assumes one latent variable.
}
\examples{
\dontrun{
data(EHR)
sample.set = EHR$sample.set
res1 <- LVS.score(X = sample.set[seq(1,50),])
}
}
\references{
Jenni Niku, Wesley Brooks, Riki Herliansyah, Francis KC Hui, Sara Taskinen, and David I Warton. Efficient estimation of generalized linear latent variable models. PloS one, 14(5), 2019.
}
