% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.curves.R
\name{multiple.pr.curve}
\alias{multiple.pr.curve}
\title{Multiple PR curves}
\usage{
multiple.pr.curve(
  disease,
  output.date = NULL,
  score.mat,
  score.names,
  group,
  group.levels,
  legend.pos = "topright",
  pairs.sub = NULL
)
}
\arguments{
\item{disease}{A string of disease name.}

\item{output.date}{An output date. The default is system date (today).}

\item{score.mat}{A resulting score matrix with multiple types of scores.}

\item{score.names}{A vector that indicates the scores of interest for comparison.}

\item{group}{A grouping vector.}

\item{group.levels}{A vector of grouping levels that reflects the severity from low to high. The typical first element represents control or healthy status.}

\item{legend.pos}{A vector of legend position. The default is "bottomright" for all plots.}

\item{pairs.sub}{An index vector for subset of group level pairs, e.g. the pairs contrasting control vs. other case stages.}
}
\value{
It produces a pdf file of plots.
}
\description{
Generate a pdf file with multiple plots for several group level contrast. Each plot consists of PR curves for different quantitative disease scores, and the corrsponding AUPRs for performance comparison.
}
\examples{
\dontrun{
multiple.pr.curve(disease = "Disease",
  score.mat = example.scores$score.mat,
  score.names = example.scores$score.names,
  group = example.scores$group,
  group.levels = example.scores$group.levels,
  legend.pos = "topright", pairs.sub = 1:5)
}
}
