% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF1.R
\name{NMF1}
\alias{NMF1}
\title{Compute Non-negative Matrix Factorization with Rank 1 Scores (NMF1s).}
\usage{
NMF1(X, p.seed = 123)
}
\arguments{
\item{X}{The original data set. It should be a matrix of numbers.}

\item{p.seed}{A random seed for model fitting.}
}
\value{
lvs The resulting LVSs for the new set Y.
}
\description{
\code{NMF1} can compute NMF1 scores with factor loadings constrained to be non-negative.
}
\examples{
\dontrun{
data(EHR)
sample.set = EHR$sample.set
res1 <- NMF1(X = sample.set[seq(1,50),])
}
}
