% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.auc.R
\name{pairwise.auc}
\alias{pairwise.auc}
\title{Pairwise Area Under the ROC Curve}
\usage{
pairwise.auc(x, g, format = "l")
}
\arguments{
\item{x}{A response vector.}

\item{g}{A grouping vector or factor.}

\item{format}{Format of result. The default is "l", long format. The other option is square format.}
}
\value{
A data frame with group 1, group 2, and the AUROC with comparison between them.
}
\description{
\code{pairwise.auc} calculates AUROC for pairwise levels of group contrast .
}
\examples{
\dontrun{
data(EHR)
x = c(rnorm(2500),runif(2500))
res = pairwise.auc(x, EHR$sample.group)
}
}
