% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.upr.R
\name{pairwise.upr}
\alias{pairwise.upr}
\title{Pairwise Area Under the PR Curve}
\usage{
pairwise.upr(x, g, format = "l")
}
\arguments{
\item{x}{A response vector.}

\item{g}{A grouping vector or factor.}

\item{format}{Format of result. The default is "l", long format. The other option is square format.}
}
\value{
A data frame with group 1, group 2, and the AUPRC with comparison between them.
}
\description{
Calculates AUPRC for pairwise group levels contrast.
}
\examples{
\dontrun{
data(EHR)
x = c(rnorm(2500),runif(2500))
res = pairwise.upr(x, EHR$sample.group)
}
}
