% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.wilcox.R
\name{pairwise.wilcox}
\alias{pairwise.wilcox}
\title{Pairwise Wilcoxon Rank Sum Tests}
\usage{
pairwise.wilcox(x, g, p.adjust.method = "bonferroni")
}
\arguments{
\item{x}{A response vector.}

\item{g}{A grouping vector or factor.}

\item{p.adjust.method}{Method for adjusting p values (can be abbreviated): "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". The default is "bonferroni".}
}
\value{
A data frame with group 1, group 2, and the Wilcoxon Rank Sum test p value of comparison between them.
}
\description{
\code{pairwise.wilcox} calculates pairwise comparisons between group levels with corrections for multiple testing.
}
\examples{
\dontrun{
data(EHR)
x = c(rnorm(2500),runif(2500))
res = pairwise.wilcox(x, EHR$sample.group)
}
}
