% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVS.R
\name{predictLVS}
\alias{predictLVS}
\title{Latent Variable Prediction}
\usage{
predictLVS(Y, model)
}
\arguments{
\item{Y}{A new data set. It should be a matrix of numbers. Default is NULL, then X will be used in prediction.}

\item{model}{A trained model.}
}
\value{
lvs The resulting LVSs for the new set Y.
}
\description{
Predicts Latent Variable Scores given a model.
}
\examples{
\dontrun{
data(EHR)
sample.set = EHR$sample.set
res1 <- predict.LVS(Y = sample.set[seq(1,50),])
}
}
