% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictQDRS.R
\name{predictQDRS}
\alias{predictQDRS}
\title{Predict Quantitative Disease Risk Score}
\usage{
predictQDRS(Y, weights)
}
\arguments{
\item{Y}{A new data set for prediction.}

\item{weights}{A matrix of weights.}
}
\value{
It returns a matrix of resulting quantitative disease risk scores.
}
\description{
Predicts Quantitative Disease Risk Scores using weights derived from training set.
}
\examples{
\dontrun{
data(EHR)
res1 <- PC(X = EHR$sample.set,
  group = EHR$sample.group,
  training = EHR$training, pc.num = 1:2)
res2 <- predictQDRS(Y = EHR$sample.set[seq(1,50),], weights = res1$weights)
}
}
