% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{rankOne.R}
\alias{rankOne.R}
\title{R Matrix with Rank One for Eigen Approach}
\usage{
rankOne.R(Qmat)
}
\arguments{
\item{Qmat}{The correlation matrix Q.}
}
\value{
\code{Rmat} An estimate of the rank-one R matrix
}
\description{
Computes an estimated rank-one matrix R with unit-norm eigenvector v. Up to a sign ambiguity, the entries of v are proportional to the balanced accuracies (the average between the sensitivity and the specificity) of the input features.
}
