% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVS.R
\name{LVS.score}
\alias{LVS.score}
\title{Compute Latent Variable Scores (LVSs).}
\usage{
LVS.score(X, Y = NULL, starting.choice = "random", p.seed = 3080)
}
\arguments{
\item{X}{The training data set. It should be a matrix of binary numbers.}

\item{Y}{The new data set. It should be a matrix of binary numbers.}

\item{starting.choice}{Starting values can be generated by fitting model without latent variables, and applying factorial analysis to residuals to get starting values for latent variables and their coefficients (starting.choice = "res"). Another options are to use zeros as a starting values (starting.choice = "zero") or initialize starting values for latent variables with (n x 1) matrix. Defaults to "random".}

\item{p.seed}{A random seed for model fitting.}
}
\value{
lvs The resulting LVSs for the new set Y.
}
\description{
\code{LVS.score} can compute Latent Variable Scores based on an unsupervised multivariate mixed model framework for multivariate bernoulli data that assumes one latent variable.
}
\examples{
\dontrun{
data(EHR)
sample.set = EHR$sample.set
res1 <- LVS.score(X = sample.set[seq(1,50),])
}
}
\references{
Jenni Niku, Wesley Brooks, Riki Herliansyah, Francis KC Hui, Sara Taskinen, and David I Warton. Efficient estimation of generalized linear latent variable models. PloS one, 14(5), 2019.
}
