% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF1.R
\name{NMF1}
\alias{NMF1}
\title{Non-negative Matrix Factorization with Rank 1 Score}
\usage{
NMF1(X, p.seed = 123)
}
\arguments{
\item{X}{The original data set. It should be a matrix of numbers.}

\item{p.seed}{A random seed for model fitting.}
}
\value{
It returns a list of following components:
\item{weights}{The NMF1 weights for input features.}
\item{scores}{The resulting NMF1 scores for the whole set.}
}
\description{
Computes NMF1 scores with factor loadings constrained to be non-negative.
}
\examples{
\dontrun{
data(EHR)
sample.set = EHR$sample.set
res1 <- NMF1(X = sample.set[seq(1,50),])
}
}
