% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.R
\name{PC}
\alias{PC}
\title{Individual Principal Components with Selected Signs}
\usage{
PC(X, group, training, pc.num)
}
\arguments{
\item{X}{The original data set that include training and test sets. It should be a matrix of numbers.}

\item{group}{A grouping vector or factor that indicate cases ("Case") and controls ("Control"). \code{NA} is allowed, and means that the observation is not used in individual PC sign determination.}

\item{training}{A logical or index vector to indicate whether the subject belongs to the training set.}

\item{pc.num}{The vector of desired individual PCs.}
}
\value{
It returns a list of following components:
\item{weights}{The selected individual PCs' weights for input features.}
\item{scores}{The resulting individual PC scores for the whole set.}
}
\description{
Computes individual PC weights and scores. The approach only requires weak labels to help select the signs of individual PCs.
}
\examples{
\dontrun{
data(EHR)
res1 <- PC(X = EHR$sample.set,
  group = EHR$sample.group,
  training = EHR$training, pc.num = 1:2)
}
}
